Rework everything to be a bit more clear
The idea with this implementation is that you create a ValenceUrlResolver with the reply from /d2l/api/versions/ (in the future, just supply a URL and this will do the call itself), then call resolver.resolve() with a ValenceRoute that has the product code, prefix, suffix, and required version (future: allow a ValenceRoute to have multiple acceptable versions). This works in theory, but is currently fairly messy.